nextbase-api/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
google-cloud-firestore
